+++ IfThisThenThat/outcomes/Outcomes.lua
@@ -3,6 +3,7 @@ local IFTTT = IFTTT
 local Outcomes = IFTTT.Outcomes or ZO_DeferredInitializingObject:Subclass()
 Outcomes.items = {
   Collectible = {},
+  CompanionCollectible = {},
 }
 
 function Outcomes:Initialize(parent)