+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2026.03.20
+- Add deliberate exclusion of Fishing
+- Add Companion Collectible as outcome
+
 ## 2026.03.19
 - Add trigger Swimming
 - Add trigger Combat - mostly working